requirements.txt
streamlit
pandas
easyocr
opencv-python-headless
Pillow
numpy
openpyxl
xlsxwriter